:rocket: Modify the logic of the run_client command
The run_client command now uses the number of days, either specified in the admin panel or entered manually through the terminal.
Priority has manual entry through the terminal. Also in the cron-command, the number of days has been changed from 160 to 60, since according to
the terms of the task, it should not exceed 3 months.